Screen: Change HiDPI display setup code in 'OpenWindow'.
If the code detects that it is running on a Retina style
display and doesn't want to expose native resolution, iow.
it has to enable the panelfitter, then set the frontend
resolution to the passed in normalized 'rect' instead of
the Screen rect, so the virtual framebuffer etc. have the
size specified by usercode, if any.

-> Maybe fixes issue #237
-> Not tested on any Retina setup yet, Screen for OSX needs rebuild.