Screen/Linux: Allow for higher calibration timing noise on pageflipped setup.
If standard (non-OpenML/non-Wayland) timestamping is used during
video refresh calibration and sync tests and we are positive that
pageflipping under our control was used, then accept a noise (stddev)
of up to 3 times the regular threshold, ie., at standard threshold
setting of 200 usecs, tolerate up to 600 usecs before sync failure.

This applies to Linux and OSX with the proprietary AMD Catalyst
driver, as we only implement GPU low level pageflip detection for
AMD gpu's on Linux and OSX+PsychtoolboxkernelDriver at the moment,
and the codepath is only executed for Catalyst, not for the open-source
drivers, where we have much better mechanisms to guarantee precise and
well defined timing.

-> This should allow Catalyst setups which work correctly but have
   noisy timing to pass the initial sync tests, and from there on
   our high precision timestamping will take care of denoising the
   timestamps.

-> Screen for 64-Bit Octave on Linux rebuilt.