Screen/Linux: Improvements to OpenML bits - Fix some checks.
The startup check to detect certain Intel ddx bugs was accidentally
disabled due to some earlier change. The check would still have
triggered on defective drivers, but at runtime, not during startup.
Revert to older "detect trouble early" behaviour.

Some minor other improvements to OpenML swap scheduling, mostly
to make the Intel ddx triple-buffering mode more happy. We don't
schedule swaps to occur in the past anymore, but make sure that
any scheduled swap is scheduled for at least the next vblank.
This because for swaps in the past, at least the intel ddx under
DRI2 + SNA + TripleBuffer on will try to optimize such a swap
away and replace it by a buffer exchange. This is probably a
OML_sync_control spec violation, but it is what it is atm. and
maybe we can get some use of the somewhat broken triple-buffering
of Intel ddx at some point if we accomodate its quirks better.

Also some updates of code comments.