Screen/Linux: Add some fixups for Intel-DDX with DRI2/SNA + TripleBuffering
These hacks make the Intel ddx 2.99.917 work under DRI2/SNA
with triple-buffering enabled - for a lenient definition of
"work". It is useable for some scenarios, hangs on others.

The ddx works perfectly fine under DRI2 with triple-buffering
disabled, just as in the past. It also works perfectly fine
under DRI3/Present with sufficiently up to date X-Server and
Mesa (XOrg 1.16.3+, Mesa 10.5+). But we'd like to get at
least some basic use out of DRI2 + triple-buffering for some
scenarios if possible, so try to accomodate some of the ddx
quirks.

The current quirks of the ddx are known:

- One must render to a framebuffer before scheduling a
  swap, or it will get optimized away into a buffer
  exchange -- chaos ensues.

- One must make sure that no more than 1 swaps are pending
  when doing a wait for swap completion, or a complete
  hang happens. Either wait for each swap, effectively
  reverting to double-buffering, or wait long enough
  for the swap chain to become idle or "single swap pending"

-> Add proper useless draw calls and wait calls to try
   to ensure above requirements are met under normal use.

-> There are various problematic use cases, but standard
   use may be able to work with these hacks. Actual timing
   tests with Datapixx et al. are tbd.

-> This commit is a stop-gap measure until i find the root
   cause of the problem in the ddx, or give up finding it.