PsychHID/OSX: Remove redundant locking from KbQueue implementation.
We used some locking leftover from the port from Linux.
This locking is not harmful but redundant, so remove it
to simplify code to simplify debugging and future work.

Apparently the "occassional hang on KbQueueRelease" but still
persists. Reason unclear, but removing useless complexity
would at least make debugging a bit more easy.

-> Rebuild and test of PsychHID pending.