PsychHID/OSX: Make KbQueues even less racy, improve error handling.
Add proper cleanup code to KbQueueCreate for all failure
cases to never leave it in an undefined state. Not that
this ever caused problems so far, but when we are at it
anyway it doesn't hurt to harden it a bit more.

More importantly try to fix further races if KbQueueCreate/Release
is executed back-to-back. The last commit didn't fix the CFRunloopStop()
race completely. We can still get into a situation where the thread is
created in KbQueueCreate, but then KbQueueRelease is executed and
calls CFRunLoopStop() before the thread managed to fully initialize
and enter its CFRunLoopRun() function. Iow., the runloop isn't yet
running at the time of the CFRunLoopStop() command and the stop
request goes into nirvana --> run loop never stops --> thread never
exits --> hang in KbQueueRelease. This is less likely to happen during
production use, but could be hit in tight Create/Release test loops
or if system timing is off a lot due to high load.

We now check in a polling loop iff the runloop reference has been
initialized and if the runloop is idle, waiting for events via
CFRunLoopIsWaiting(). The implementation is not exactly textbook,
but we can afford a bit of inefficiency due to polling here, as
create/release are not executed in timing sensitive bits of experiment
scripts. The poor man's method saves us from quite a bit complexity,
as doing this efficiently and correct with Apple's api's would be quite
some nightmare, likely to introduce new bugs an traps on its own.

-> Code changed, but recompile and test pending.