Screen/Linux/X11: In GetFlipInfo, replace "s" flag with "?"
At the time we process the swap events on GLX, we simply don't know
(anymore) if they were associated with a vsync'ed or non-sync'ed
swapbuffers op, and there isn't a simple way to track this - at
least none worth the effort for this minor information.

Therefore always return "?" for the vsync'ed flag instead of
"s" for synced or "_" for non-synced.