PsychPortAudio: Try to avoid selecting HDMI/DP outputs as default soundcard.
Since many gpu's come with builtin sound chips that output the
audio stream via DisplayPort or HDMI to a capable display, we
can get into the situation that those gpu's somehow get assigned
as default soundcard and the user hears the sound of silence without
any indication why.

Try to detect such an unfortunate default assignment by a blacklist
of name snippets like "HDMI", "hdmi", "isplay" ( for D[isplay]Port)
and try to reassign a better soundcard if this is the case. If no
better card can be found, reset to PortAudio's default assignment.

I don't know if this is a good idea or not. Only time will tell...

-> PsychPortAudio for 64-Bit Octave on Linux rebuilt and tested.

-> On my NVidia GeForce GT 330M with HDMI audio output, the internal
   Intel HDA analog sound output gets assigned by Linux as the default
   card, the NVidia only comes in second. Iow., the problem doesn't
   exist on my machine, so it is not clear how prominent this would
   be on other users machines. Another problem is that selection of,
   e.g., the digital output of a builtin "real" soundcard is not
   neccessarily a better choice, not knowing what the user really
   wants and has connected.