Screen/Wayland: Suppress "non vsync present" warning for opaque non-fs windows.
We know opaque windowed windows can end up in gpu hardware overlays
on Wayland's Weston reference compositor as of at least version 1.8,
and those are always reported non-vsync presented due to limitations
and bugs in Weston's current experimental implementation, so suppress
the "no vsync" warning clutter for the moment.

Once Weston gains proper atomic planes support, reenable the warning
to report unexpected problems. Note that this may not (only) be a
limitation of Weston at the moment, but a limitation of the Linux
kernels planes implementation, which itself is still experimental
work in progress with known limitations. Overlays may need fixing
in Weston and Linux to finally be useable in a research grade fashion.