Screen/Wayland: Implement scheduling compensation for Waylands scheduling.
Try to target the earliest point in time inside the video refresh cycle
directly before the target stimulus onset frame, so the compositor will
lock onto the start vblank of the following target frame. For this we
need to find out from given targetTime in which frame the targetTime
is located, then find the start time of active scanout of that frame
and return that as corrected targetTime.

Background:

Some (most?) Wayland compositors will define some composition deadline
within their current frame. Presentation requests from clients arriving
after that deadline will not be processed for the current frame, but delayed
to the next frame, in order to leave sufficient headroom in the current
frame for the time consuming rendering needed for desktop composition of
new content for the upcoming frame, so they won't skip that frames vblank
deadline and cause skipped frames. At least Weston as of 1.8 has such a
deadline defined as 'repaint-window' msecs before end of current frame, with
a default value of 7 msecs, ie., requests arriving < 7 msecs before end of
a frame will be delayed by a full frame. Weston may get a dynamic deadline
if my experimental patches get accepted, to provide a shorter cutoff if
possible, or the strategy may change to something like "unredirect fullscreen
windows" like on X11, or the future presentation_queue extension may deal with
this elegantly, but atm. for Weston 1.8 this problem will exist. The strategy
of other Wayland compositors (KDE, GNOME etc.) is not yet known.

In any case, targetting the earliest possible swap submission time for given
targetTime makes us as robust as possible against different compositors strategies.

-> Screen.mex for Wayland on 64-Bit Octave/Linux rebuilt.

-> Tested against Weston 1.8-rc with my experimental patches applied,
   with/without hardware overlays, fullscreen/non-fs, with/without
   a manually set extra delay. Seems to work well.