Screen/Wayland: Implement conversion of input event timestamps to GetSecs time.
Wayland input event timestamps are only 32-Bit uint with msecs
resolution, as provided by the underlying libinput library. We
need GetSecs time, so add a remap helper function. 32-Bit msecs
timestamps would wrap around in less than 2 months of system
uptime which is too low, so we first recover 64-Bit millisecond
timestamps from the delivered input event timestamps before
converting into double precision GetSecs time.

We can do the recovery of full 64-Bit range because we know that
libinput asks the Linux kernels evdev device to return input events
which are time stamped with CLOCK_MONOTONIC nanosecond precision
timestamps, but then throws away all that nice high precision and
quantizes to msecs timestamps of which only the lowermost 32-Bits
are returned. Anyway, we know the timebase and can use a clock query
to it for the current CLOCK_MONOTONIC time to recover the upper 32-Bit
of the timestamps + use some wraparound handling, knowing that the
input event won't have happened in the future, and likely won't have
happened 2 months in the past, relative to the query time.

Note to self: Try to prepare patches to extend libinput upstream
for microsecond precision 64-Bit timestamps. Won't be useable by
Waylands current protocol, but might be useful for our own low-level
queries to input devices in the future, e.g., for a DRM/KMS backend
or raw device passthrough.

-> Screen for Wayland+Octave+Linux rebuilt and tested.
-> Works nicely.