Screen/Wayland: Allow identity clut loading via colord profiling inhibit.
Use colord's cd_device_profiling_inhibit function to setup
identity luts for LoadIdentityClut() functionality if on
Wayland + colord and getenv('PSYCH_USE_COLORD_IDENTITYLUT', '1').

In theory this is an elegant solution to the identity clut
problem. In practice it doesn't work at all on Weston :(
Do i misunderstand the api, or is this just another bug
of Westons colord backend?

-> Screen for 64-Bit Octave on Linux+Wayland rebuilt.