Screen/Wayland: Make ReadNormalizedGammatable more "robust".
Another hack: If we can't get a gamma table from the icc
profile, we use tone curve reconstruction to synthesize one.
The table can be outside the actual valid range 0.0 - 1.0
for real gamma tables. Therefore normalize the values, so
the max value always maps to 1.0 and we avoid out of range
errors from LoadNormalizedGammaTable if the read out lut
gets written later again.

-> Screen for 64-Bit Octave on Linux + Wayland rebuilt and tested.