OSX: Add AutoBrightness function.
AutoBrightness() allows to get/set the “automatically adjust display
brightness” setting on OS X. It uses AppleScript to do this.
The method simulates user input via accessibility api’s and has a good
chance of breaking in future OS X versions, so we only put it into the
PsychContributed folder, marking it as “not officially supported, no
guarantee it will work”.

Unfortunately, research showed that there doesn’t seem to be a cleaner
or more robust way to control this setting - no public programming
api’s, no low level IORegistry or IOKit methods, so this potentially
fragile hack seems to be the best we can do.

Contributed by Denis Pelli.