Screen/Linux/X11: Prevent crash at Screen reload on multi-x-screen with buggy Mesa.
Current Mesa versions have a bug in both the nouveau and radeon winsys drivers
which would cause a crash if we'd close the X connection here and later Screen()
gets reloaded again within the same Octave/Matlab session. By avoiding to close the
connection here we can avoid this crash bug, at the minor expense of leaking a bit of
memory and X resources.

The bug has been already fixed by myself upstream for nouveau, with patches
for the AMD side of things reviewed and pending inclusion, so we just use this
workaround on Mesa versions known to contain the bug. Currently these would be
stable Mesa 10.6.x and earlier stable versions, but backports of these fixes
are expected within the next weeks at least for Mesa 10.6 and probably 10.5.

-> Add detection logic for running on Mesa and for specific Mesa version,
   accessible by window glue and screen glue.

-> Window glue sets the detected version during onscreen window open.
-> Screen glue checks for buggy Mesa versions on multi-x-screen during
   shutdown to avoid XCloseDisplay() when it would be potentially hazardous.