Screen/Linux: Add some missing new dynamic library dependencies on XCB.
For our DRI3/Present detection/handling code we need libX11-XCB,
libxcb and libxcb-dri3. These were included in the Octave build
by luck, but are missing in the Matlab build. Add them properly
to the makefiles for both Matlab and Octave to fix the Matlab
link failure and make dependencies clear on the Octave side.

-> linuxmakeit64.m and linuxmakeitoctave.m updated.
-> Screen mex for 64-Bit Octave/Matlab on Linux rebuilt and tested.