Update and extend conversions between degrees and mm of retina
Main work was adding routines RetinalEccentricityMMToDegrees and DegreesToRetinalEccentricityMM, which implement formulae from the literature that take the shape and optics of the eye into account.

Doing this also led to some minor changes in DegreesToRetinalMM and RetinalMMToDegrees.  I changed the default behavior to make these implement the small angle linear approximation, rather than the tangent plane trig approximation.  For the linear approximation, I did this so that the two exactly self invert.  Also extended the comments to try to explain everything clearly.

Signed-off-by: David Brainard <brainard@psych.upenn.edu>