NetStation: Improve and consolidate various improvements.
"I worked on a new and much improved version of the Netstation script,
which I attach. There have been several problems with the one included
in the distribution now. It should be replaced with the one here, which
has also been tested by EGI, the company who produces the Net Station
software for EEG acquisition. The new version includes all the additions
and improvements that various people contributed to it during the years,
and it is now free from the bugs that also sneaked into it. It should
work both on PCs and on Macs."

Improvements or consolidation of others improvements contributed
by Gergely Csibra and tested by EGI - the manufacturer of the EGI
Netstation system.