Screen: Add new optional 'ignoreErrors' flag to 'LoadNormalizedGammaTable'
This new flag if set to 1 will cause Screen() to not abort a script
if loading a gamma table fails in immediate mode. Instead it will
just print a warning to the console and return a 'failed' status
code as new 2nd optional return argument. This allows iterative
tweaking of gamma tables until they get accepted by the OS.

This is relevant to work around gamma table limitations imposed
by MS-Windows for exotic applications. See forum message 19342
for reference.

-> Tested on Linux. Needs recompile and testing on OSX and Windows.