Screen: Make FTGL plugin the default text renderer on OSX.
Make the switch. The default text renderer 1 is now the FTGL
text rendering plugin not only on Linux, as in the past, but
also on OSX. This because we want to get rid of that awful
nightmare that is Apples native text rendering technology.

We try to make a smooth transition: Renderer 1 uses the FTGL
plugin, which should work on OSX. If that fails for some reason
we fall back to text renderer 0, which on OSX maps to the old
legacy Apple CoreText text renderer. Users still can explicitely
select renderer 0 via Screen('Preference','TextRenderer', 0) if
they need Apples horror cabinet for some reason.

Lets see if anybody seriously complains and rather fix the FTGL
renderer in that case. Given that FTGL outperforms the Apple tech
in every single aspect i'm aware of, this shouldn't be a problem
but a net improvement for users and for long term maintenance.

If everything goes well and Apple should decide to break their renderer
the umpteenth time with a future OSX release, we can simply rip the
old legacy renderer out and be done with it.

-> Screen for 64-Bit Octave on Linux rebuilt and tested.
-> Rebuild and test on OSX pending.
-> TextToStufColorMismatchTest.m adapted to the new situation that
   CoreText is now renderer 0 and not renderer 1 anymore.