Screen: Improve Screen('Close', window);
Make it do more sensible cleanup, mostly matching the behaviour of
Screen(‘CloseAll’). When the last onscreen window gets closed, we now
call ScreenCloseAllWindows to get the full Screen(‘CloseAll’) cleanup.

When a single onscreen window is closed, make sure to release its
associated screen if the onscreen window was the last one on that
screen and the screen was captured.