Screen: Add 'detachFromMouse' support to SetMouse().
Allow to detach mouse cursor movements from mouse movements, so the
cursor stays locked in place and can only be manually moved via
SetMouse(). SetMouse() and Screen(‘SetMouseHelper’) have a new optional
‘detachFromMouse’ flag.

This is currently only implemented for OS X. We could do this on
Linux/X11 by setting mouse slave pointers to floating, or on Windows
maybe via the ClipCursor() function, although that function seems to
have a lot of problematic failure modes, according to the internet.

-> Screen for 64-Bit Octave on OS X rebuilt and tested.
-> SetMouse() updated.