Screen/OSX: Fix hostname display in Screen('Computer')
Switch the C-String target character encoding from the modern UTF-8 to
the old-school ASCII. Apparently Matlab on OS X can’t deal with host
names that contain apostrophe characters. If we assign
computer.machineName as UTF-8 with such a hostname, character trash is
returned instead of the apostrophe. I could not reproduce this problem
on OS X 10.10.3 with Octave 3.8, so maybe it is only Matlab on OS X
which is too primitive to deal with such things. Anyway, changed the
encoding, hope it will help. Test on Matlab is pending, on Octave
everything is still fine.

-> Screen.mex for 64-Bit Octave on OS X rebuilt and tested.
-> See forum message number 19263 for reference.