Screen/Linux: Make ext_buffer_age warning less chatty.
Some versions of the proprietary Nvidia driver seem to misreport buffer
age via the EXT_buffer_age query. This leads to false positive warnings
about non-doublebuffered flips (single-buffered or triple-buffered) and
timestamping failure.

Knowing that ext_buffer_age is not a perfectly reliable way to check
for double-buffer failure on different Nvidia drivers, lets tone the
warnings down a bit from “Your screwed!” to a suggestion to run some
diagnostic checks. Also make it a one-time warning per onscreen window
per session, so on properly working setups which misreport buffer age
we don’t get timing problems due to the flood of printed warning output
to the Matlab window - ie. avoid turning the warning message into a
self-fulfilling prophecy.

-> Ideally i should find out which range of drivers is affected by the
false positives, and just blacklist those, or when it happens, but i
don’t have a large enough sample of Nvidia cards and thereby sufficient
variability in driver versions and gpu models to really do this.
However, if timing would be really broken, PTB should detect this
indirectly by the other startup tests and runtime tests, just as it
does on OS X or Windows, so we can treat ext_buffer_age as just another
debugging hint or cue for troubleshooting. If other sync failures or
runtime warnings happen at the same time as this one-time warning, that
would be a good indicator of trouble.