New display correction method: DisplayUndistortionLabRiggerMouseStim.m
This is a template based on undistortion code for mouse visual
stimuli from LabRigger:

http://labrigger.com/blog/2012/03/06/mouse-visual-stim/comment-page-1/

This file needs to be improved for general usability by
an actual user of the code. It is just meant as a starting
point/example on how to integrate such calibrations into
PTB's fast image processing pipeline.