Factor out OS X Octave rpath handling into extra function file.
osxsetoctaverpath.m now contains the routine for setting @rpath entries
in Octave mex files for OS X. The code was previously contained in
osxmaketoctave.m. This way it can also be used by datapixxmakemex to
edit the rpath in Datapixx.mex for OS X Octave, something we so far
failed to do.