Oculus VR: Fixup distortion parameters better.
Now we first query the PsychOculusVRCore driver for recommended rendebuffer
sizes for a given FOV (or default FOV as recommended by the Oculus runtime),
and use it as our virtual framebuffer size for the PanelFitter as before.

But then we allocate the inputBufferFBO/processedDrawBufferFBO's which
feed into the undistortion rendering to be twice the width of the actual
framebuffer (aka HMD panel width) and the height of the actual framebuffer.
This way we get an inputbuffer that is big enough to cover all typical
use cases, and the maximum the current Screen imaging pipeline can do.
Then we feed this actual size of the input buffers into the Oculus runtime
for computation of the distortion meshes, texture scale/offset factors and
other rendering parameters, so these correspond to the actual real sizes
of our distortion image left/right input buffers. Result seems to be more
symmetric/plausible scaling and offset factors for texture sampling, and
possibly improved image quality (could be a placebo though).

To allow this, the core driver has now separate split up functions for
getting client renderbuffer size and undistortion parameters, so we can
do this in two phases.

Downside is now a bigger input buffer to sample for undistortion and
to upscale via the panelfitter from input client buffers. Iow. another
performance reduction.

But first make it work correctly, then make it fast.

-> New baseline.
-> Mex file updated, PsychImaging updated, PsychOculusVR.m updated.