PsychImaging: Add support for running VR HMD's on secondary outputs of a screen.
This is mostly to allow quick debugging of VR HMD code when one is
too lazy to reconfigure the display system, e.g., to create a separate
X-Screen for the HMD on Linux/X11.

PsychImaging detects if one tries to display the VR HMD's onscreen window
on a PTB screen which has multiple active video outputs attached. If so,
it warns about the massive performance implications (decreased framerate,
increased latency, bad stimulus onset timing and unreliable timestamping)
and gives recommendations for an optimized setup. Then it tries to find
the output corresponding to the HMD and generates a proper onscreen window
'rect' to position the onscreen window exclusively on that output.

-> Convenience rules.