Screen: Add new infoType 7 query to 'GetWindowInfo'
'infoType' 7 does the same as the default 'infoType' 0, but
omits PsychSetDrawingTarget() ie. doesn't activate the windows
OpenGL context and doesn't set it up as drawingtarget. Also avoids
returning info which needs the OpenGL context of the window.

This allows to query a subset of all winfo information without
touching OpenGL state. This is mostly needed for VR HMD support,
specifically our implementation of Oculus Timewarp, as we need
to Screen('GetWindowInfo') from within our image post-processing
to get the timestamp of last vblank in order to do predictive warping.

Without avoiding PsychSetDrawingTarget we end with corrupted
imaging pipeline FBO bindings and a black screen.

-> Screen for 64-Bit Octave on Linux rebuilt.