PsychOculusVRCore: Add new method 'GetEyePose' for querying per eye tracking state.
PsychOculusVRCore('GetEyePose') queries the runtime for the predicted headpose
at the time when the rendered image for a given eye renderpass will display.

It translates the input 'renderPass' into an eyeIndex, then queries for
that eyeIndex, using the SDK recommended eye render order to optimize
rendering order for the update order of the displays scanout.

The returned eye pose is actually a HMD pose or head pose, so it needs
to be translated into actual predicted eye pose by application of the
eye shift vectors, e.g., as provided by PsychOculusVR('GetEyeShiftVector')
for a given eye.

-> PsychOculusVRCore.mex for 64-Bit Octave on Linux rebuilt.

-> In theory this should provide better prediction and thereby better
   aligned rendering of per eye views, as the HMD's IMU head tracker
   gets resampled for a better pose prediction at each invocation of
   'GetEyePose'. Afaik the IMU updates at 1000 Hz, ie. once every msec,
   so any per-eye render that lasts longer than 1 msec will result in
   a updated hardware tracked pose, which is good. In practice, with
   the n=1 test scene we have, there isn't any perceptible difference.