VR support: New helper function eyePoseToCameraMatrix()
This converts an eye pose vector into a 4x4 matrix that
describes absolute position and orientation of the camera.

This is the counterpart to eyePoseToCameraGLModelviewMatrix(),
in that it compute the forward transform, not the inverse.

Use this for renderers which want camera position, instead
of inverse camera position for rendering, e.g., Horde3D.