Oculus VR: Add support for the Rift DK2 builtin latency tester.
A new function PsychOculusVRCore(‘LatencyTester’) allows to setup and
query the latency tester for the draw color to be used for latency
testing, and for the results of the latency test.

PsychOculusVR.m contains code to query the to-be-used color for the
latency test marker pixel for each rendered frame, then to draw the
marker pixel at the top-left corner of the framebuffer.

-> Allows to measure latency via ground-truth from a sensor in the
device. More specific: Similar to what the VPixx devices like Datapixx
do, the DK2 can parse the input video stream for certain marker pixels
in certain marker colors in the top-left corner of the scanout aka
start of active scanout of a new video frame. Reception of the marker
pixel gets timestamped, transmitted to the runtime and there magic
happens, which is very similar to what i implemented in Psychtoolbox
for VPixx devices, to calculate things like latency between start of
render and start of scanout of rendered frame.

-> PsychOculusVRCore.mex for Octave and Matlab on OS X rebuilt.
-> PsychOculusVR.m updated.
-> VRHMDDemo1.m contains some temporary test code for this stuff.
-> Needs testing on a sufficiently powerful GPU to see if it actually
works.