VR support: Add more subfunctions to PsychVRHMD and PsychOculusVR.
PsychVRHMD should "support" a certain set of standardized functions
which are useful cross-vendor, cross-device. We "implement" much of
these by only providing a help text for them. The actual implementations
of the functions are proxied/dispatched to functions of the same name
and syntax in the vendor specific drivers, e.g., PsychOculusVR.

PsychOculusVR received help texts for a couple of subfunctions where
help texts were missing.