PsychVRHMD, PsychOculusVR: Fail gracefully in 'AutoSetupHMD'.
If no connection to the underlying hardware driver(s) is possible,
e.g., because the machine doesn't have the required runtime libraries
installed, then simply fail with a message saying so and return an
empty handle, so caller can deal with this, maybe using some fallback
path.

VRHMDDemo1 adapted to take advantage of this.