VR Support: New and improved subfunctions for PsychVRHMD / PsychOculusVR.
The 'PrepareRender' function is rewritten. It now only supports
reqflags = 1, which is also the default. In this mode it returns
both forward and inverse camera matrices for both eyes for a given
tracked head pose. It takes a targetTime for predicting the headpose
and uses the expected midpoint of video display for the to-be-rendered
frame as a good baseline value if targetTime is omitted. The tracked
headpose now gets premultiplied with a user provided global pose matrix,
to allow globally positioning and orienting of the observer.

Additionally the 'GetEyePose' method from Oculus is implemented in the
same fashion as 'PrepareRender', just for individual eyes. I finally
found a situation where per-eye individual renderpasses really help:
Fast head movements will show less prediction error and a better VR
experience with per eye rendering, instead of just taking the values
from 'PrepareRender'. As not all drivers will be able to support
'GetEyePose' in a meaningful way - so far only the Oculus driver for
Rift DK1/DK2 is - there exists a new 'GetInfo' subfield to signal to
usercode if use of 'GetEyePose' will be beneficial for quality, or if
it would just be a waste of processing time on the given HMD. All
drivers are required to implement the method nonetheless. On a incapable
driver, the correct behaviour can be emulated by simply caching the
tracking results of the mandatory 'PrepareRender' subfunction internally,
then using those cached results in the 'GetEyePose' pseudo implementation
to calculate per eye pose from that cached head pose. Not meaningfully
enhancing quality, but keeps code working.

The new helper method PsychGetPositionYawMatrix() can now be used
to position the observer in the world, and then the generated matrix
getting passed to PsychVRHMD('PrepareRender' / 'GetEyePose') to apply
head tracking relative to the global pose, so we have the needed
ingredients for moving in a large scene and implementing basic game
logic.

-> VRHMDDemo1.m udpated to use the new approach. Also cleanups and
   refinements to the movement "game logic".