Commit full folder with TestConfigurations for timing tests.
Just to get rid of that annoying setup step whenever i want
to do timing tests. 4.5 MB are not that bad for a bit of
convenience.

One still needs to create the TimingtestResults/ subfolder
as Psychtoolbox/PsychTests/TimingtestResults/ though, so
the test script knows where to store its results by default.