PsychPortAudio/Windows: Add basic Octave-4 support.
PsychPortAudio needs special treatment atm. We can't use our own
libportaudioX64.dll as it is not compatible with Octave-4's Mingw32
build system. Instead, as a stop-gap measure, we use the portaudio dll
bundled with Octave-4. This works for the basics, but lacks our special
features: DirectInputMonitoring and ASIO support. Without ASIO this is
very basic and not really research grade, but as a stepping stone should
do.

-> Add shim implementation of missing functions to PsychPortAudio.c for
now. And some special case to InitializePsychSound.
-> PsychPortAudio.m
ex for 64-Bit Octave-4 on Windows built.