Screen/Windows/Octave: Don't attach thread input on Octave.
For some reason the Octave-4 GUI does not like it at all if we attach
our thread input to the QT based GUIs input. It ends with a GUI hang at
the end of each Screen() session. Therefore don't do this on Octave, but
only for Matlab. Without it, not only are the hangs gone, but all
Keyboard and Mouse input seems to work normally with Octave, as tested
with GetChar and GetCharTest et al., KbCheckDemo and KbQueueDemo, as
well as MouseTraceDemo et al.