PsychOculusVR: Protect against failed loading of wrong mex file.
On Octave, a missing PsychOculusVRCore.mex file for the target OS and
machine architecture can cause Octave to try to load the mex file for a
different OS+arch, e.g., if Windows build is missing, it tries to load
from the Linux folder. This ends badly.

Improve error handling for the 'Supported' function, so proper
'unsupported' status is returned and HMD autosetup skips Oculus runtime
drivers.