Windows: Also implant and call PsychStartup() into Octaves startup.
Implant a call to PsychStartup() into Octave's .octaverc startup file
for Octave on Microsoft Windows, so the path to the GStreamer runtime
directory is properly setup. This is crucial as on Octave we can't
delay-load the GStreamer dll's as with Matlab. Octave always forces
resolution of all link time dependencies at mex file load time. This
means GStreamer is not a soft requirement only when using multi-media
functions, but a hard requirement for Screen() to work at all, even
without use of GStreamer multi-media functionality.