Linux: Update build scripts and tools to support Octave-4 on Linux.
linuxmakeitoctave3 and datapixxmakemex now also copy each mex
file into the corresponding Octave4LinuxXXX subfolder and run
striplibsfrommexfile with new optional 3rd argument set to 1.

This will cause striplibsfrommexfile to rename the dynamic
dependency on liboctinterp.so.2 into liboctinterp.so.3, as
that is the new name of liboctinterpreter on Octave-4. This
way the linker can link our old mex files which were built
against Octave-3 against the new libraries of Octave-4.

This should work - and seems to work in practice - as Octave-4's
C MEX API and ABI has not changed wrt. Octave-3. Fingers crossed!