Screen: Add new ConserveVRAM setting kPsychAssumeGfxCapVCGood.
This enforces setting the kPsychGfxCapVCGoo property for a window,
so drawing of textures with color clamping disabled no longer enforces
use of the special GLSL texture filter shaders, as is normally the case.

This as a workaround for the bug found by Peter Scarfe on OSX 10.11
with AMD graphics hardware, where texture sampling with bilinear filtering
by use of our filter shader gives very inaccurate results, apparently due
to a new driver bug in Apples creature. This is a potential workaround, not
yet tested and confirmed to actually help. But it would be the only way we
could work around this, so it is this or nothing.

-> help ConserveVRAMSettings updated.
-> Screen for 64-Bit Octaves on Linux rebuilt and tested.