Screen('DrawDots'): Add shader-based fallback for point smoothing.
If 'dot_type' 3 is requested, or if the gfx-driver is known to not
support GL_POINT_SMOOTH, as is currently the case with the FOSS
Mesa drivers on Linux for at least Intel and AMD (but not nouveau
for NVidia), then use a GLSL shader based fallback implementation,
based on point sprites and the gl_PointSize GLSL version 1.10
variable. We use texunit 1's texture coordinate set to pass in
fragment coordinates relative to the GL_POINTS center, as auto-
generated via point sprite extension. The fragment shader uses
these together with the known point size to calculate location of
an output fragment in pixels relative to the point center, and thereby
the radius of the fragment. Fragments outside the radius + 0.5 are
discarded, fragments inside the points radius - 0.5 are drawn at
full alpha, fragments in the border radius - 0.5 to radius + 0.5
are faded out by using alpha-values that decrease from 1.0 to 0.0
over that 1 pixel length unit. gl_PointSize is set in the vertex
shader as passed in via a uniform to control the to be rasterized
area of the GL_POINT.

-> Seems to give results of comparable quality of typical point
   smoothing as tested on Linux + Intel HD Ironlake.

-> We use the point sprite extension instead of gl_PointCoord builtin
   variable as this allows us to still work on GLSL versions 1.0 and 1.1,
   whereas the builtin variable would need GLSL version 1.2 or later,
   iow. it would require OpenGL 2.1 or later, instead of also working
   with OpenGL 1.5 and 2.0.