Screen('DrawDots'): Speedup drawing of dots of different sizes.
New dot_type setting 4 allows to draw square dots as with
default setting 0, but uses our GLSL shaders to do so. This
allows to pass in dot sizes as texture coordinates and set them
in the vertex shader via gl_PointSize. Iow. we can use the same
optimization as with shader based round dot drawing in order to
submit all to be drawn dots in one draw call instead of having to
submit one draw call for each dot. This is vastly more efficient
and removes the last remaining slow path on GLSL capable hardware.

-> All modes of operation of DrawDots should now be fast.
-> Tested on Intel HD ironlake under Linux, speedup is high.
-> Screen.mex for 64-Bit Octaves under Linux rebuilt.