Improve PsychLinuxConfiguration() further for ease of use.
The script records all yes/no answers from the user. If
setup fails due to lack of privileges, it outputs a command
line with the recorded answers, to be run in a suitable admin
context. When called that way, it will use the recorded answers
which are passed in as first argument and skip asking the same
yes/no questions again.

Additionally the script now offers to add the current user,
and optionally any number of existing additional users, to the
psychtoolbox and dialout Unix user groups, so the user can skip
that manual setup step.

Also upgrading the psychtoolbox limits file is now possible
if the current distro comes with a more recent file than the
one installed, in case we need to tweak realtime scheduling
permissions or settings in the future.

Also some output messages slightly improved.